requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
nomic==1.1.7
numpy==1.24.3
pydantic==2.4.2
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pgvector==0.2.3
python-dotenv==1.0.0 